+++ module/member/alert.inc.php
@@ -1,16 +1,18 @@
 <?php 
 defined('IN_DESTOON') or exit('Access Denied');
 login();
+$MG['biz'] or dalert(lang('message->without_permission_and_upgrade'), 'goback');
 require DT_ROOT.'/module/'.$module.'/common.inc.php';
 $MOD['alertid'] or message($L['feature_close']);
 $MG['alert_limit'] > -1 or dalert(lang('message->without_permission_and_upgrade'), 'goback');
+$menu_id = 2;
 $mids = array();
 $tmp = explode('|', $MOD['alertid']);
 foreach($tmp as $v) {
 	if($v > 4 && isset($MODULE[$v])) $mids[] = $v;
 }
 require DT_ROOT.'/include/post.func.php';
-require MD_ROOT.'/alert.class.php';
+require DT_ROOT.'/module/'.$module.'/alert.class.php';
 $do = new alert();
 
 switch($action) {
@@ -40,11 +42,14 @@ switch($action) {
 					$$v = '';
 				}
 				$mid = $_mid;
+				$head_title = $L['alert_add_title'];
 			} else {
-				$action = 'choose';
+				dheader('?action=list');
 			}
-			$head_title = $L['alert_add_title'];
 		}
+	break;
+	case 'list':
+		$head_title = $L['alert_add_title'];
 	break;
 	case 'edit':
 		$itemid or message();
@@ -68,10 +73,12 @@ switch($action) {
 	break;
 	case 'delete':
 		$itemid or message();
-		$do->itemid = $itemid;
-		$r = $do->get_one();
-		if(!$r || $r['username'] != $_username) message();
-		$do->delete($itemid);
+		$itemids = is_array($itemid) ? $itemid : array($itemid);
+		foreach($itemids as $itemid) {
+			$do->itemid = $itemid;
+			$item = $do->get_one();
+			if($item && $item['username'] == $_username) $do->delete($itemid);
+		}
 		dmsg($L['op_del_success'], $forward);
 	break;
 	default:
@@ -80,14 +87,10 @@ switch($action) {
 		$condition = "username='$_username' AND status=$status";
 		$lists = $do->get_list($condition);
 		if($lists) {
-			$tmp = $MOD['linkurl'];
 			foreach($lists as $k=>$v) {
-				if($v['catid']) {
-					$lists[$k]['cate'] = cat_pos(get_cat($v['catid']), '-', 1);
-				}
+				if($v['catid']) $lists[$k]['cate'] = $DT_PC ? cat_pos(get_cat($v['catid']), '-', 1) : strip_tags(cat_pos(get_cat($v['catid']), '-'));
 				if($v['email'] != $_email) $db->query("UPDATE {$DT_PRE}alert SET email='$_email' WHERE itemid=$v[itemid]");
 			}
-			$MOD['linkurl'] = $tmp;
 		}
 		$head_title = $L['alert_title'];
 	break;
@@ -100,5 +103,18 @@ for($i = 2; $i < 4; $i++) {
 	$limit_used += $r['num'];
 }
 $limit_free = $MG['alert_limit'] && $MG['alert_limit'] > $limit_used ? $MG['alert_limit'] - $limit_used : 0;
+if($DT_PC) {
+	//
+} else {
+	$foot = '';
+	if($action == 'edit' || $action == 'list') {
+		$back_link = '?action=index';
+	} else if ($action == 'add') {
+		$back_link = '?action=list';
+	} else {
+		$pages = mobile_pages($items, $page, $pagesize);
+		$back_link = ($kw || $page > 1) ? '?action=index' : 'biz.php';
+	}
+}
 include template('alert', $module);
 ?>
